requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
flake8
black
yapf
python-gitlab
twine
six==1.15.0
numpy==1.18.5
pandas==1.3.5
pandas-gbq==0.16.0
pandas-profiling==3.1.0
scikit-learn==0.22
scikit-image==0.19.1
joblib
memoized-property
mlflow
s3fs
gcsfs==2022.7.1
google-cloud-storage==2.4.0
termcolor
